in_red_fs: more work on reinstate failed drive(s), laid out the function stubs across modules/files. Will fill up with code to implement them next.
