+++ fs/registry_copier.go
@@ -1,0 +1,10 @@
+package fs
+
+import (
+	"context"
+)
+
+func (r *registryOnDisk) CopyToPassiveFolders(ctx context.Context) error {
+
+	return nil
+}
