+++ fs/replication_tracker.go
@@ -21,6 +21,10 @@ type replicationTrackedDetails struct {
 	// If true, folder as specified in storesBaseFolders[0] will be the active folder,
 	// otherwise the 2nd folder, as specified in storesBaseFolders[1].
 	ActiveFolderToggler bool
+
+	// If true, Transactions should log commit changes so they can be used for "fast forward" functionality
+	// of ReinstateFailedDrives functionality.
+	LogCommitChanges bool
 }
 
 func (a replicationTrackedDetails) isEqual(b replicationTrackedDetails) bool {
@@ -102,66 +106,6 @@ func (r *replicationTracker) HandleReplicationRelatedError(ctx context.Context, 
 			log.Error(fmt.Sprintf("failover to folder %s failed, details: %v", r.getPassiveBaseFolder(), err.Error()))
 		}
 	}
-}
-
-// ReinstateFailedDrives can be invoked after replacing the failed drives so they can be synchronized with
-// the current Active targets' Stores' contents.
-//
-// Details:
-//   - Tell transactions to start logging commit changes
-//   - Copy the registries & storeRepositories files to the new drives
-//   - Once done, switch over the entire cluster into "InDeltaSync" mode
-//   - Using all the transactions' commit logs' generated while doing the 2nd step above,
-//     fast forward registries & storeRepositories files
-//   - Update the L2 cache copy of global replication status to turn back on, the replication
-//     to the passive targets
-//   - Turn off the "InDeltaSync" mode to switch over to normal mode
-func (r *replicationTracker) ReinstateFailedDrives(ctx context.Context) error {
-
-	if err := r.startLoggingCommitChanges(ctx); err != nil {
-		return err
-	}
-	tr := sop.NewTaskRunner(ctx, -1)
-	tr.Go(func() error {
-		return r.copyStoreRepositories(tr.GetContext())
-	})
-	tr.Go(func() error {
-		return r.copyRegistries(tr.GetContext())
-	})
-	if err := tr.Wait(); err != nil {
-		return err
-	}
-	if err := r.fastForward(ctx); err != nil {
-		return err
-	}
-	if err := r.turnOnReplication(ctx); err != nil {
-		return err
-	}
-	// Check fast forward log one last time.
-	return r.fastForward(ctx)
-}
-
-func (r *replicationTracker) startLoggingCommitChanges(ctx context.Context) error {
-	return nil
-}
-func (r *replicationTracker) copyRegistries(ctx context.Context) error {
-
-	return nil
-}
-func (r *replicationTracker) copyStoreRepositories(ctx context.Context) error {
-	return nil
-}
-func (r *replicationTracker) fastForward(ctx context.Context) error {
-	// TODO: apply commit logs, since commit logging got started, one after another until caught up.
-	return r.syncWithL2Cache(ctx, true)
-}
-func (r *replicationTracker) turnOnReplication(ctx context.Context) error {
-	globalReplicationDetails.FailedToReplicate = false
-	globalReplicationDetails.ActiveFolderToggler = !globalReplicationDetails.ActiveFolderToggler
-
-	// Update the replication status details.
-	r.writeReplicationStatus(r.formatActiveFolderEntity(replicationStatusFilename))
-	return r.syncWithL2Cache(ctx, true)
 }
 
 func (r *replicationTracker) handleFailedToReplicate(ctx context.Context) {
