+++ fs/store_repository_copier.go
@@ -1,0 +1,10 @@
+package fs
+
+import (
+	"context"
+)
+
+func (sr *StoreRepository) CopyToPassiveFolders(ctx context.Context) error {
+
+	return nil
+}
